+++ src/snowflake/snowpark/async_job.py
@@ -248,6 +248,14 @@ class AsyncJob:
         status = self._session._conn._conn.get_query_status(self.query_id)
         return self._session._conn._conn.is_an_error(status)
 
+    def status(self) -> str:
+        """
+        Returns a string representing the current status of the query.
+        (e.g., "RUNNING", "SUCCESS", "FAILED_WITH_ERROR", "ABORTING", etc.)
+        """
+        status = self._session._conn._conn.get_query_status(self.query_id)
+        return status.name
+
     def cancel(self) -> None:
         """Cancels the query associated with this instance."""
         # stop and cancel current query id